server/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
pydantic==2.6.3
starlette==0.37.2
alembic==1.13.1
SQLAlchemy==2.0.24
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.9
email_validator==2.2.0